School Module added but not workinng.............plase assit
There is a Capitla Simple issue but can't identify
