+++ src/main/java/lk/successStudent/studentManagement/asset/school/dao/schoolDao.java
@@ -1,11 +1,11 @@
 package lk.successStudent.studentManagement.asset.school.dao;
 
 
-import lk.successStudent.studentManagement.asset.school.entity.School;
+import lk.successStudent.studentManagement.asset.school.entity.school;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SchoolDao extends JpaRepository<School, Integer> {
+public interface schoolDao extends JpaRepository<school, Integer> {
 
 }
