+++ src/main/java/lk/successStudent/studentManagement/asset/school/entity/school.java
@@ -1,6 +1,7 @@
 package lk.successStudent.studentManagement.asset.school.entity;
 
 
+
 import com.fasterxml.jackson.annotation.JsonFilter;
 import lk.successStudent.studentManagement.util.audit.AuditEntity;
 import lombok.AllArgsConstructor;
@@ -8,13 +9,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import javax.persistence.Entity;
+
+@Entity
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@JsonFilter( "School" )
-public class School extends AuditEntity {
+@JsonFilter("School")
+public class school extends AuditEntity {
+    private String schoolName;
+    private String schoolCode;
+    private String district;
 
-    private String schoolId;
 
-    private String name;}
+
+}
